+++ engine/src/core/Trade.h
@@ -5,7 +5,7 @@ struct Trade
 {
     uint64_t tradeId;
     uint64_t securityId;
-    double price;
+    uint64_t price;
     uint32_t quantity;
     uint64_t aggressingOrderId;
     uint64_t restingOrderId;
